Refactor the helper project for host API tests
This makes the project modular - with ability to easily add tests for other APIs (the upcomming hostpolicy ones).
This is a pure refactoring, the only functional change is the addition of logging the full path to the hostfxr being used by the tests.
